test(app): colocate message action integration
