+++ src/app/message_actions.rs
@@ -266,4 +266,6 @@ impl App<'_> {
 }
 
 #[cfg(test)]
+mod integration_tests;
+#[cfg(test)]
 mod tests;
